feat: added cpu resources and finailized initial resource action implementation
This commit finalizes the implementation for cpu resources
gpu resources now follow step policy rather than having seperate
increment decrement functions
gpu resources are now bounds checked
fixed gpu resources not being included in json ouput
gpu resources assumed to be nvidia
removed gpu_model option for now
overall consolidated the codebase to use simmilar conventions

Note: theres a somewhat significant naming issue here, to "request" a
gpu your actually supposed to put it in "limits" this is a weird api
call on the kube side, we could put it in request part of the json but
its bad form and im not 100% sure it works. so our gpu resources being
stored in the requests struct of the container mismatches the json
structure. that being said it is a resource request so its not entirely
off base to put it there.